requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=9.0.2
pytest-django>=4.11.1
pytest-cov>=7.0.0
coverage>=7.13.0

# Code quality
black>=25.12.0
isort>=7.0.0
flake8>=7.3.0
pyright>=1.1.407

# Documentation
sphinx>=8.2.0,<9
sphinx-rtd-theme>=3.0.2
